livekit-self-hosted/agent/requirements.txt
livekit-agents>=1.0.0
livekit-plugins-silero>=1.0.0
livekit-api>=1.0.0
python-dotenv>=1.0.0
